requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
numpy>=1.24
pandas>=2.0
pyyaml>=6.0

# ── Computer Vision & Pose Estimation ─────────────────────────────────────────
opencv-python>=4.8
mediapipe>=0.10

# ── Deep Learning ─────────────────────────────────────────────────────────────
torch>=2.1
torchvision>=0.16

# ── Graph Neural Networks ─────────────────────────────────────────────────────
torch-geometric>=2.4
# torch-scatter and torch-sparse are installed separately — see README.md

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.7
seaborn>=0.13
networkx>=3.2          # skeleton graph visualisation in notebooks

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0
ipykernel>=6.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4
